# Proposed fix for #40