This PR edits thread_safe_queue to have more functionality and serve as a priority queue for the threads. 

At the end of each thread execution, the threads bump themselves up to the top of the priority list for work assignment.