@DeveloperPaul123 2023-07-03
Overall this looks really good. It seems that this has a significant performance impact on the thread pool as well but I'm skeptical on the numbers I'm seeing. Here are the new benchmark results with MSVC


| relative |               ms/op |                op/s |    err% |     total | matrix multiplication 8x8
|---------:|--------------------:|--------------------:|--------:|----------:|:--------------------------
|   100.0% |               73.49 |               13.61 |    0.8% |     18.22 | `dp::thread_pool - std::function`
|   105.5% |               69.62 |               14.36 |    0.3% |     16.94 | `dp::thread_pool - std::move_only_function`
|    98.0% |               74.99 |               13.34 |    1.1% |     18.22 | `dp::thread_pool - fu2::unique_function`
|    39.7% |              185.12 |                5.40 |    0.4% |     44.72 | `BS::thread_pool`
|    50.7% |              144.83 |                6.90 |    0.7% |     34.89 | `task_thread_pool`
|    40.2% |              182.84 |                5.47 |    0.5% |     44.23 | `riften::Thiefpool`

| relative |               ms/op |                op/s |    err% |     total | matrix multiplication 64x64
|---------:|--------------------:|--------------------:|--------:|----------:|:----------------------------
|   100.0% |               58.25 |               17.17 |    4.4% |     14.69 | `dp::thread_pool - std::function`
|   108.8% |               53.55 |               18.68 |    0.7% |     13.01 | `dp::thread_pool - std::move_only_function`
|   101.3% |               57.53 |               17.38 |    0.9% |     13.94 | `dp::thread_pool - fu2::unique_function`
|    40.3% |              144.52 |                6.92 |    0.3% |     34.92 | `BS::thread_pool`
|    45.2% |              128.80 |                7.76 |    0.2% |     31.15 | `task_thread_pool`
|    42.4% |              137.50 |                7.27 |    0.3% |     33.40 | `riften::Thiefpool`

| relative |               ms/op |                op/s |    err% |     total | matrix multiplication 256x256
|---------:|--------------------:|--------------------:|--------:|----------:|:------------------------------
|   100.0% |               47.36 |               21.11 |    2.6% |     11.54 | `dp::thread_pool - std::function`
|   106.5% |               44.49 |               22.48 |    0.2% |     10.74 | `dp::thread_pool - std::move_only_function`
|   100.1% |               47.31 |               21.14 |    0.2% |     11.46 | `dp::thread_pool - fu2::unique_function`
|    47.0% |              100.80 |                9.92 |    0.3% |     24.42 | `BS::thread_pool`
|    51.1% |               92.77 |               10.78 |    0.4% |     22.41 | `task_thread_pool`
|    46.3% |              102.20 |                9.78 |    0.7% |     24.99 | `riften::Thiefpool`

| relative |               ms/op |                op/s |    err% |     total | matrix multiplication 512x512
|---------:|--------------------:|--------------------:|--------:|----------:|:------------------------------
|   100.0% |               38.02 |               26.30 |    0.3% |      9.20 | `dp::thread_pool - std::function`
|   134.7% |               28.23 |               35.42 |    2.3% |      6.85 | `dp::thread_pool - std::move_only_function`
|    97.3% |               39.08 |               25.59 |    0.6% |      9.50 | `dp::thread_pool - fu2::unique_function`
|    49.5% |               76.79 |               13.02 |    0.2% |     18.57 | `BS::thread_pool`
|    53.1% |               71.65 |               13.96 |    0.2% |     17.35 | `task_thread_pool`
|    48.6% |               78.29 |               12.77 |    0.4% |     18.91 | `riften::Thiefpool`

| relative |               ms/op |                op/s |    err% |     total | matrix multiplication 1024x1024
|---------:|--------------------:|--------------------:|--------:|----------:|:--------------------------------
|   100.0% |               42.87 |               23.33 |    1.9% |     10.32 | `dp::thread_pool - std::function`
|    99.9% |               42.91 |               23.31 |    2.2% |     10.30 | `dp::thread_pool - std::move_only_function`
|   101.9% |               42.08 |               23.76 |    1.4% |     10.21 | `dp::thread_pool - fu2::unique_function`
|    73.0% |               58.70 |               17.03 |    0.6% |     14.22 | `BS::thread_pool`
|    77.5% |               55.32 |               18.08 |    0.7% |     13.45 | `task_thread_pool`
|    99.5% |               43.08 |               23.21 |    2.2% |     10.47 | `riften::Thiefpool`

@jtd-formlabs Do you have any input or insights on what could be causing this large of a performance uplift? Previous benchmarks showed my library edging out some other popular libraries but now it blows them out of the water. I'm not mad at that, but I'm wondering if there is something I'm missing here...