@jtd-formlabs 2023-07-03
This is really interesting! I'll run your benchmarks on my Ubuntu system to get some additional metrics as well. 

One theory is that this new scheduling system is relying on the work stealing process much less, as it will always defer to a thread that is ready for work first so there is less delay.

It may be worth adding some code to before and after this PR to see how many workloads are stolen during the benchmarks.