@DeveloperPaul123 2023-07-03
Yes I agree with your comments. This should result in the threads always having work to do in general and having to steal less (if at all). 

I'm curious to see the numbers on Ubuntu as well as I haven't tried running benchmarks there yet.