@jtd-formlabs 2023-07-03
So I ran two benchmarks, one with `pyperf system` and one without:

WITH:
```

| relative |               ms/op |                op/s |    err% |     total | matrix multiplication 8x8
|---------:|--------------------:|--------------------:|--------:|----------:|:--------------------------
|   100.0% |              182.55 |                5.48 |    0.4% |     32.67 | `dp::thread_pool - std::function`
|    99.0% |              184.43 |                5.42 |    0.3% |     33.04 | `dp::thread_pool - std::move_only_function`
|    98.9% |              184.60 |                5.42 |    0.5% |     33.13 | `dp::thread_pool - fu2::unique_function`
|    74.9% |              243.85 |                4.10 |    0.3% |     44.00 | `BS::thread_pool`
|    56.0% |              326.01 |                3.07 |    0.5% |     58.38 | `task_thread_pool`
|   132.7% |              137.60 |                7.27 |    1.6% |     25.69 | `riften::Thiefpool`

| relative |               ms/op |                op/s |    err% |     total | matrix multiplication 64x64
|---------:|--------------------:|--------------------:|--------:|----------:|:----------------------------
|   100.0% |              159.72 |                6.26 |    0.6% |     28.60 | `dp::thread_pool - std::function`
|    96.7% |              165.12 |                6.06 |    0.3% |     29.62 | `dp::thread_pool - std::move_only_function`
|   100.5% |              158.87 |                6.29 |    0.4% |     28.42 | `dp::thread_pool - fu2::unique_function`
|    82.0% |              194.82 |                5.13 |    0.4% |     35.09 | `BS::thread_pool`
|    63.3% |              252.48 |                3.96 |    0.4% |     45.24 | `task_thread_pool`
|   105.8% |              150.96 |                6.62 |    4.1% |     27.27 | `riften::Thiefpool`

| relative |               ms/op |                op/s |    err% |     total | matrix multiplication 256x256
|---------:|--------------------:|--------------------:|--------:|----------:|:------------------------------
|   100.0% |              118.77 |                8.42 |    0.2% |     21.26 | `dp::thread_pool - std::function`
|    92.9% |              127.82 |                7.82 |    0.7% |     22.89 | `dp::thread_pool - std::move_only_function`
|    99.0% |              119.94 |                8.34 |    0.7% |     21.49 | `dp::thread_pool - fu2::unique_function`
|    81.2% |              146.29 |                6.84 |    1.0% |     26.19 | `BS::thread_pool`
|    66.0% |              180.02 |                5.55 |    0.1% |     32.26 | `task_thread_pool`
|    88.6% |              134.01 |                7.46 |    1.5% |     24.10 | `riften::Thiefpool`

| relative |               ms/op |                op/s |    err% |     total | matrix multiplication 512x512
|---------:|--------------------:|--------------------:|--------:|----------:|:------------------------------
|   100.0% |               99.88 |               10.01 |    1.3% |     17.86 | `dp::thread_pool - std::function`
|    99.9% |              100.00 |               10.00 |    0.3% |     17.93 | `dp::thread_pool - std::move_only_function`
|   101.0% |               98.93 |               10.11 |    0.3% |     17.75 | `dp::thread_pool - fu2::unique_function`
|    85.7% |              116.59 |                8.58 |    0.7% |     20.96 | `BS::thread_pool`
|    76.1% |              131.28 |                7.62 |    0.8% |     23.60 | `task_thread_pool`
|   100.0% |               99.90 |               10.01 |    4.1% |     18.02 | `riften::Thiefpool`

| relative |               ms/op |                op/s |    err% |     total | matrix multiplication 1024x1024
|---------:|--------------------:|--------------------:|--------:|----------:|:--------------------------------
|   100.0% |               78.76 |               12.70 |    1.5% |     14.20 | `dp::thread_pool - std::function`
|    95.5% |               82.48 |               12.12 |    1.0% |     14.78 | `dp::thread_pool - std::move_only_function`
|   100.1% |               78.70 |               12.71 |    1.1% |     14.08 | `dp::thread_pool - fu2::unique_function`
|    78.6% |              100.23 |                9.98 |    1.3% |     18.05 | `BS::thread_pool`
|    77.3% |              101.88 |                9.82 |    0.5% |     18.29 | `task_thread_pool`
|    88.8% |               88.66 |               11.28 |    3.3% |     15.83 | `riften::Thiefpool`
```

WITHOUT:
```
Warning, results might be unstable:
* CPU frequency scaling enabled: CPU 0 between 400.0 and 4,700.0 MHz
* CPU governor is 'powersave' but should be 'performance'
* Turbo is enabled, CPU frequency will fluctuate

Recommendations
* Use 'pyperf system tune' before benchmarking. See https://github.com/psf/pyperf

| relative |               ms/op |                op/s |    err% |     total | matrix multiplication 8x8
|---------:|--------------------:|--------------------:|--------:|----------:|:--------------------------
|   100.0% |              114.40 |                8.74 |    2.0% |     20.56 | `dp::thread_pool - std::function`
|    92.6% |              123.52 |                8.10 |    0.5% |     22.06 | `dp::thread_pool - std::move_only_function`
|    92.7% |              123.39 |                8.10 |    0.7% |     22.20 | `dp::thread_pool - fu2::unique_function`
|    70.6% |              162.04 |                6.17 |    0.3% |     28.88 | `BS::thread_pool`
|    53.2% |              214.96 |                4.65 |    0.6% |     38.59 | `task_thread_pool`
|   114.4% |              100.03 |               10.00 |    0.3% |     18.20 | `riften::Thiefpool`

| relative |               ms/op |                op/s |    err% |     total | matrix multiplication 64x64
|---------:|--------------------:|--------------------:|--------:|----------:|:----------------------------
|   100.0% |              114.62 |                8.72 |    1.0% |     20.52 | `dp::thread_pool - std::function`
|   104.1% |              110.14 |                9.08 |    0.8% |     19.79 | `dp::thread_pool - std::move_only_function`
|   106.4% |              107.77 |                9.28 |    1.0% |     19.31 | `dp::thread_pool - fu2::unique_function`
|    86.4% |              132.71 |                7.54 |    1.5% |     23.65 | `BS::thread_pool`
|    68.8% |              166.60 |                6.00 |    0.7% |     29.88 | `task_thread_pool`
|   112.9% |              101.49 |                9.85 |    1.6% |     18.09 | `riften::Thiefpool`

| relative |               ms/op |                op/s |    err% |     total | matrix multiplication 256x256
|---------:|--------------------:|--------------------:|--------:|----------:|:------------------------------
|   100.0% |               82.04 |               12.19 |    0.6% |     14.73 | `dp::thread_pool - std::function`
|    97.6% |               84.06 |               11.90 |    0.6% |     14.99 | `dp::thread_pool - std::move_only_function`
|   101.0% |               81.26 |               12.31 |    2.1% |     14.56 | `dp::thread_pool - fu2::unique_function`
|    81.7% |              100.41 |                9.96 |    0.6% |     17.96 | `BS::thread_pool`
|    68.2% |              120.34 |                8.31 |    0.4% |     21.49 | `task_thread_pool`
|    86.0% |               95.37 |               10.49 |    1.3% |     17.16 | `riften::Thiefpool`

| relative |               ms/op |                op/s |    err% |     total | matrix multiplication 512x512
|---------:|--------------------:|--------------------:|--------:|----------:|:------------------------------
|   100.0% |               68.46 |               14.61 |    0.6% |     12.26 | `dp::thread_pool - std::function`
|   100.9% |               67.83 |               14.74 |    1.0% |     12.21 | `dp::thread_pool - std::move_only_function`
|   100.1% |               68.37 |               14.63 |    0.3% |     12.26 | `dp::thread_pool - fu2::unique_function`
|    84.3% |               81.25 |               12.31 |    0.7% |     14.59 | `BS::thread_pool`
|    72.1% |               94.99 |               10.53 |    0.6% |     17.00 | `task_thread_pool`
|    92.2% |               74.27 |               13.46 |    1.6% |     13.36 | `riften::Thiefpool`

| relative |               ms/op |                op/s |    err% |     total | matrix multiplication 1024x1024
|---------:|--------------------:|--------------------:|--------:|----------:|:--------------------------------
|   100.0% |               59.33 |               16.86 |    0.6% |     10.66 | `dp::thread_pool - std::function`
|    95.8% |               61.94 |               16.14 |    1.1% |     11.14 | `dp::thread_pool - std::move_only_function`
|    97.9% |               60.61 |               16.50 |    1.0% |     10.82 | `dp::thread_pool - fu2::unique_function`
|    81.9% |               72.44 |               13.80 |    1.0% |     13.04 | `BS::thread_pool`
|    71.4% |               83.11 |               12.03 |    3.3% |     15.51 | `task_thread_pool`
|    92.2% |               64.37 |               15.54 |    1.0% |     11.47 | `riften::Thiefpool`
```

This was on a laptop running Ubuntu 20.04 with the power cable plugged in. 20 core system with 32 gb of ram on an Intel processor.